djangoreact/backend/requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
black==21.11b1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
environ==1.0
et-xmlfile==1.1.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0